feat(game): add .start() and .playGame() methods
The `.start()` method is what tells the game to prepare to be started. The server will then wait for all players to be connected before calling `.playGame()`, which will return a Promise chain that runs the game logic. The Promise chain completing is what signifies the end of the game.